Fix up some issues in the InterfaceQuery and InterfaceRecord classes:
Define a constant key for "interface-name" and use that in place of the incorrect RECORD_SERVICE_NAME.

Take out support for storing subnets in the interface-address array; no code seems to use this.

Make the code that processes the interface address arrays in the JSON handle the data found in the sLS database; in particular, don't choke if there's no '/' character.

Adjust various getters and setters to correctly handle the List<String> data in the JSON values, rather than Strings or ints, so we can actually read existing data.

Make getters for numeric quantities gracefully handle malformed values.

Fixes issue #27.